Chatbot Perpustakaan Esa Unggul Citra Raya/requirements.txt
flask
gunicorn
fuzzywuzzy
python-Levenshtein
mysql-connector-python
gtts